Parser update (part 2)
Parser class contains static method 'run' which takes Scanner as input, and returns an array list of Channels

+++ La Serena/src/Note.java
@@ -1,7 +1,7 @@
 public class Note {
-    public final float value;
+    public final double value;
     public final int length;
-    Note(float value, int length)
+    Note(double value, int length)
     {
         this.value = value;
         this.length = length;
